docs/cookbook/requirements.txt
# TKE Workshop Cookbook Python Dependencies

# 腾讯云 SDK
tencentcloud-sdk-python>=3.0.1000
tencentcloud-sdk-python-common>=3.0.1000
tencentcloud-sdk-python-tke>=3.0.1000
tencentcloud-sdk-python-cvm>=3.0.1000

# Kubernetes Python Client
kubernetes>=28.1.0
pyyaml>=6.0

# 配置文件解析
pyyaml>=6.0

# 日志和工具
colorlog>=6.8.0
rich>=13.7.0
click>=8.1.7

# HTTP 请求
requests>=2.31.0

# 进度条
tqdm>=4.66.0

# 时间处理
python-dateutil>=2.8.2

# 表格输出
tabulate>=0.9.0

# 异步支持 (可选)
aiohttp>=3.9.0
asyncio>=3.4.3
